+++ src/main/java/org/seedstack/jdbc/JdbcExceptionHandler.java
@@ -5,9 +5,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-/*
- * Creation : 18 févr. 2015
- */
 package org.seedstack.jdbc;
 
 import org.seedstack.seed.transaction.spi.ExceptionHandler;