+++ src/main/java/org/seedstack/jdbc/internal/DataSourceDefinition.java
@@ -12,51 +12,21 @@ import org.seedstack.jdbc.spi.DataSourceProvider;
 
 import javax.sql.DataSource;
 
-/**
- * Holds all the objects associated to a JDBC data source.
- *
- * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
- */
 class DataSourceDefinition {
-
     private final String name;
-
+    private final DataSource dataSource;
     private Class<? extends JdbcExceptionHandler> jdbcExceptionHandler;
-
     private DataSourceProvider dataSourceProvider;
 
-    private DataSource dataSource;
-
-    /**
-     * Constructor.
-     *
-     * @param name the data source name
-     */
-    DataSourceDefinition(String name) {
-        this.name = name;
-    }
-
-    /**
-     * Constructor.
-     *
-     * @param name       the data source name
-     * @param dataSource the dataSource
-     */
     DataSourceDefinition(String name, DataSource dataSource) {
         this.name = name;
         this.dataSource = dataSource;
     }
 
-    /**
-     * @return the data source name
-     */
     String getName() {
         return name;
     }
 
-    /**
-     * @return the JDBC exception handler
-     */
     Class<? extends JdbcExceptionHandler> getJdbcExceptionHandler() {
         return jdbcExceptionHandler;
     }
@@ -75,9 +45,5 @@ class DataSourceDefinition {
 
     DataSource getDataSource() {
         return dataSource;
-    }
-
-    void setDataSource(DataSource dataSource) {
-        this.dataSource = dataSource;
     }
 }