+++ src/main/java/org/seedstack/jdbc/internal/datasource/C3p0DataSourceProvider.java
@@ -5,15 +5,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-/*
- * Creation : 18 févr. 2015
- */
 package org.seedstack.jdbc.internal.datasource;
 
 import com.mchange.v2.c3p0.ComboPooledDataSource;
-import org.seedstack.seed.SeedException;
 import org.seedstack.jdbc.internal.JdbcErrorCode;
 import org.seedstack.jdbc.spi.DataSourceProvider;
+import org.seedstack.seed.SeedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -21,35 +18,34 @@ import javax.sql.DataSource;
 import java.beans.PropertyVetoException;
 import java.util.Properties;
 
-/**
- * Datasource provider for C3p0.
- *
- * @author yves.dautremay@mpsa.com
- */
 public class C3p0DataSourceProvider implements DataSourceProvider {
     private static final Logger LOGGER = LoggerFactory.getLogger(C3p0DataSourceProvider.class);
 
     @Override
-    public DataSource provide(String driverClass, String url, String user, String password, Properties jdbcProperties) {
-        ComboPooledDataSource cpds = new ComboPooledDataSource();
+    public DataSource provide(String driverClassName, String url, String user, String password, Properties jdbcProperties) {
+        ComboPooledDataSource comboPooledDataSource = new ComboPooledDataSource();
+
         try {
-            cpds.setDriverClass(driverClass);
-            cpds.setJdbcUrl(url);
-            cpds.setUser(user);
-            cpds.setPassword(password);
-            cpds.setProperties(jdbcProperties);
-            return cpds;
+            comboPooledDataSource.setDriverClass(driverClassName);
+            comboPooledDataSource.setJdbcUrl(url);
+            comboPooledDataSource.setProperties(jdbcProperties);
+            if (user != null) {
+                comboPooledDataSource.setUser(user);
+            }
+            if (password != null) {
+                comboPooledDataSource.setPassword(password);
+            }
         } catch (PropertyVetoException e) {
-            throw SeedException.wrap(e, JdbcErrorCode.WRONG_JDBC_DRIVER);
+            throw SeedException.wrap(e, JdbcErrorCode.UNABLE_TO_PROVIDE_DATASOURCE);
         }
+
+        return comboPooledDataSource;
     }
 
     @Override
     public void close(DataSource dataSource) {
         try {
-            if (dataSource instanceof ComboPooledDataSource) {
-                ((ComboPooledDataSource) dataSource).close();
-            }
+            ((ComboPooledDataSource) dataSource).close();
         } catch (Exception e) {
             LOGGER.warn("Unable to close datasource", e);
         }