+++ src/main/java/org/seedstack/jdbc/internal/datasource/PlainDataSource.java
@@ -5,49 +5,34 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-/*
- * Creation : 18 févr. 2015
- */
 package org.seedstack.jdbc.internal.datasource;
-
-import org.seedstack.seed.SeedException;
-import org.seedstack.jdbc.internal.JdbcErrorCode;
 
 import javax.sql.DataSource;
 import java.io.PrintWriter;
 import java.sql.Connection;
-import java.sql.DriverManager;
+import java.sql.Driver;
 import java.sql.SQLException;
 import java.sql.SQLFeatureNotSupportedException;
 import java.util.Properties;
 import java.util.logging.Logger;
 
-/**
- * A simple datasource that will take its connection directly from the driver. No pooling or anything.
- *
- * @author yves.dautremay@mpsa.com
- */
-public class PlainDataSource implements DataSource {
-
+class PlainDataSource implements DataSource {
+    private final Driver driver;
+    private final String url;
+    private final Properties properties;
+    private PrintWriter writer = new PrintWriter(System.out);
     private int timeout;
 
-    private PrintWriter writer = new PrintWriter(System.out);
-
-    private String url;
-
-    private Properties properties;
-
-    public PlainDataSource(String driver, String url, String user, String password, Properties jdbcProperties) {
-        try {
-            Class.forName(driver);
-        } catch (ClassNotFoundException e) {
-            throw SeedException.wrap(e, JdbcErrorCode.WRONG_JDBC_DRIVER);
-        }
-
+    PlainDataSource(Driver driver, String url, Properties properties, String user, String password) {
+        this.driver = driver;
         this.url = url;
-        jdbcProperties.put("user", user);
-        jdbcProperties.put("password", password);
-        this.properties = jdbcProperties;
+        this.properties = properties;
+        if (user != null) {
+            properties.put("user", user);
+        }
+        if (password != null) {
+            properties.put("password", password);
+        }
     }
 
     @Override
@@ -70,6 +55,17 @@ public class PlainDataSource implements DataSource {
         return timeout;
     }
 
+    @Override
+    public Logger getParentLogger() throws SQLFeatureNotSupportedException {
+        // not using JUL
+        throw new SQLFeatureNotSupportedException();
+    }
+
+    @Override
+    public boolean isWrapperFor(Class<?> iface) throws SQLException {
+        return iface.isInstance(this);
+    }
+
     @SuppressWarnings("unchecked")
     @Override
     public <T> T unwrap(Class<T> iface) throws SQLException {
@@ -79,23 +75,16 @@ public class PlainDataSource implements DataSource {
         throw new SQLException("Wrapped DataSource is not an instance of " + iface);
     }
 
-    public Logger getParentLogger() throws SQLFeatureNotSupportedException {
-        throw new SQLFeatureNotSupportedException();
-    }
-
-    @Override
-    public boolean isWrapperFor(Class<?> iface) throws SQLException {
-        return iface.isInstance(this);
-    }
-
     @Override
     public Connection getConnection() throws SQLException {
-        return DriverManager.getConnection(url, properties);
+        return driver.connect(this.url, this.properties);
     }
 
     @Override
     public Connection getConnection(String username, String password) throws SQLException {
-        throw new UnsupportedOperationException("Not supported by PlainDataSource");
+        Properties properties = new Properties(this.properties);
+        properties.put("user", username);
+        properties.put("password", password);
+        return driver.connect(this.url, properties);
     }
-
 }